fix: toContractor method has been moved to ContractorMapper
